Variable rearrangement.
RestApplicationPort ExpectedData Environments from mock.openstack.controller.ts to testrest.datasource.ts.
