+++ app/waf/test/acceptance/ping.controller.acceptance.ts
@@ -21,15 +21,15 @@ import {
   teardownApplication,
   TestingApplication,
   setupRestAppAndClient,
-  RestApplicationPort,
   setupEnvs,
   teardownRestAppAndClient,
   teardownEnvs,
-} from '../helpers/test-helper';
+} from '../helpers/testsetup-helper';
+import {MockKeyStoneController} from '../fixtures/controllers/mocks/mock.openstack.controller';
 import {
+  RestApplicationPort,
   ExpectedData,
-  MockKeyStoneController,
-} from '../fixtures/controllers/mocks/mock.openstack.controller';
+} from '../fixtures/datasources/testrest.datasource';
 
 const prefix = '/adcaas/v1';
