+++ app/waf/test/acceptance/sequence.acceptance.ts
@@ -17,21 +17,23 @@
 import {
   TestingApplication,
   setupRestAppAndClient,
-  RestApplicationPort,
   teardownRestAppAndClient,
   setupEnvs,
   teardownEnvs,
-} from '../helpers/test-helper';
+} from '../helpers/testsetup-helper';
 import {
   MockKeyStoneController,
   OSShouldResponseWith,
-  ExpectedData,
 } from '../fixtures/controllers/mocks/mock.openstack.controller';
 import {MockSelfTestController} from '../fixtures/controllers/mocks/mock.selftest.controller';
 import {Client, expect} from '@loopback/testlab';
 import {OpenStackComponent} from '../../src/components';
 import {stubLogger, restoreLogger} from '../helpers/logging.helpers';
 import {MySequence} from '../../src/sequence';
+import {
+  RestApplicationPort,
+  ExpectedData,
+} from '../fixtures/datasources/testrest.datasource';
 
 describe('openstack.identity.test', () => {
   let mockKeystoneApp: TestingApplication;
